Refactoring without errors
