+++ src/components/shared/DrinksPage/Category/index.js
@@ -1,0 +1,1 @@
+export { default } from './Category';
