+++ src/components/shared/DrinksPage/Drinks/index.js
@@ -1,0 +1,1 @@
+export { default } from './Drinks';
