+++ src/components/shared/DrinksPage/DrinksCard/index.js
@@ -1,0 +1,1 @@
+export { default } from './DrinksCard';
