+++ src/components/shared/DrinksPage/DrinksList/index.js
@@ -1,0 +1,1 @@
+export { default } from './DrinksList';
