+++ src/components/shared/DrinksPage/Filter/index.js
@@ -1,0 +1,1 @@
+export { default } from './Filter';
