+++ src/components/shared/DrinksPage/Ingredients/index.js
@@ -1,0 +1,1 @@
+export { default } from './Ingredients';
