--- src/components/shared/DrinksPage/categories.json
@@ -1,13 +1,0 @@
-[
-  "Ordinary Drink",
-  "Cocktail",
-  "Shake",
-  "Other/Unknown",
-  "Cocoa",
-  "Shot",
-  "Coffee/Tea",
-  "Homemade Liqueur",
-  "Punch/Party Drink",
-  "Beer",
-  "Soft Drink"
-]
