--- src/components/shared/DrinksPage/redux/store.js
@@ -1,10 +1,0 @@
-// import { configureStore } from '@reduxjs/toolkit';
-// import { filterSlice } from './filterSlice';
-// import { recipetsSlice } from './recipetsSlice';
-
-// export const store = configureStore({
-//   reducer: {
-//     valueFilter: filterSlice.reducer,
-//     dataRecipets: recipetsSlice.reducer,
-//   },
-// });
