+++ src/components/shared/api/DrinkPage/getCategories.js
